Rename to pytoyoda and update everything accordingly
* Also updated some workflows to use better pypi features,
which should unblock poetry > 1
* Still needs some more fixes to README.md

Fixes #2
