+++ .github/workflows/build.yml
@@ -29,8 +29,8 @@ jobs:
           python-version: "${{ matrix.python-version }}"
       - name: "⚙️ Install Poetry"
         uses: "abatilo/actions-poetry@v4.0.0"
-        with:
-          poetry-version: 1.5.1
+        # Automatic installation of plugins fails due to https://github.com/python-poetry/poetry/issues/10028
+        # so we still do a `self add`
       - name: "⚙️ Install dependencies"
         run: |
           poetry install --without dev
