+++ .github/workflows/linting.yml
@@ -16,8 +16,6 @@ jobs:
           python-version: 3.9
       - name: "⚙️ Install Poetry"
         uses: "abatilo/actions-poetry@v4.0.0"
-        with:
-          poetry-version: 1.5.1
       - name: "⚙️ Install dependencies"
         run: "poetry install"
       - name: "🚀 Run pre-commit on all files"
