+++ .github/workflows/python-publish.yml
@@ -6,17 +6,29 @@ name: "Upload Python Package to PyPi"
 on:
   release:
     types: ["published"]
+
 jobs:
-  build:
-    runs-on: "ubuntu-latest"
+  pypi-publish:
+    name: upload release to PyPI
+    runs-on: ubuntu-latest
+    environment: pypi
+    permissions:
+      id-token: write
     steps:
       - name: "⤵️ Check out code from GitHub"
         uses: "actions/checkout@v4"
         with:
           fetch-depth: 0
           fetch-tags: true
-      - name: "🚀 Build package and publish to pypi"
-        uses: "JRubics/poetry-publish@v1.16"
-        with:
-          pypi_token: "${{ secrets.PYPI_PASSWORD }}"
-          plugins: "poetry-dynamic-versioning[plugin]"
+
+      # Automatic installation of plugins fails due to https://github.com/python-poetry/poetry/issues/10028
+      # so we still do a `self add`
+      - name: "⚙️ Install dependencies"
+        run: |
+          poetry install --without dev
+          poetry self add "poetry-dynamic-versioning[plugin]"
+      - name: "🚀 Test package building"
+        run: "poetry build"
+
+      - name: Publish package distributions to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
