+++ .github/workflows/testing.yml
@@ -27,14 +27,12 @@ jobs:
           python-version: "${{ matrix.python-version }}"
       - name: "⚙️ Install Poetry"
         uses: "abatilo/actions-poetry@v4.0.0"
-        with:
-          poetry-version: 1.5.1
       - name: "⚙️ Install dependencies"
         run: "poetry install"
       - name: "🚀 Run pytest with coverage"
         run: |
           poetry run pip install coverage
-          poetry run coverage run --source=mytoyota/ -m pytest
+          poetry run coverage run --source=pytoyoda/ -m pytest
           poetry run coverage xml
       - name: "⬆️ Upload Coverage to Codecov"
         uses: "codecov/codecov-action@v5"
