--- mytoyota/utils/__init__.py
@@ -1,1 +1,0 @@
-"""Utils for mytoyota lib."""
