+++ pytoyoda/__init__.py
@@ -1,6 +1,6 @@
 """Toyota Connected Services Client."""
 from importlib_metadata import version
 
-from mytoyota.client import MyT  # noqa
+from pytoyoda.client import MyT  # noqa
 
 __version__ = version(__name__)
