+++ pytoyoda/models/endpoints/account.py
@@ -5,8 +5,8 @@ from uuid import UUID
 
 from pydantic.v1 import Field
 
-from mytoyota.models.endpoints.common import StatusModel
-from mytoyota.utils.models import CustomBaseModel
+from pytoyoda.models.endpoints.common import StatusModel
+from pytoyoda.utils.models import CustomBaseModel
 
 
 class _TermsActivityModel(CustomBaseModel):
