+++ pytoyoda/models/endpoints/common.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Union
 
 from pydantic.v1 import Field
 
-from mytoyota.utils.models import CustomBaseModel
+from pytoyoda.utils.models import CustomBaseModel
 
 
 class UnitValueModel(CustomBaseModel):
