+++ pytoyoda/models/endpoints/electric.py
@@ -4,8 +4,8 @@ from typing import Optional
 
 from pydantic.v1 import Field
 
-from mytoyota.models.endpoints.common import StatusModel, UnitValueModel
-from mytoyota.utils.models import CustomBaseModel
+from pytoyoda.models.endpoints.common import StatusModel, UnitValueModel
+from pytoyoda.utils.models import CustomBaseModel
 
 
 class ElectricStatusModel(CustomBaseModel):
