+++ pytoyoda/models/endpoints/trips.py
@@ -7,9 +7,9 @@ from uuid import UUID
 
 from pydantic.v1 import Field
 
-from mytoyota.models.endpoints.common import StatusModel
-from mytoyota.utils.helpers import add_with_none
-from mytoyota.utils.models import CustomBaseModel
+from pytoyoda.models.endpoints.common import StatusModel
+from pytoyoda.utils.helpers import add_with_none
+from pytoyoda.utils.models import CustomBaseModel
 
 
 class _SummaryBaseModel(CustomBaseModel):
