+++ tests/unit_tests/test_utils/test_helpers.py
@@ -1,7 +1,7 @@
 """Test Helper Utils."""
 import pytest
 
-from mytoyota.utils.helpers import add_with_none
+from pytoyoda.utils.helpers import add_with_none
 
 
 # Parametrized test for happy path with various realistic test values
